refactor(): CONFIG_MAP_KEY으로 네이밍 변경
