+++ src/modules/cloud-config/cloud-config.controller.ts
@@ -1,4 +1,5 @@
 import { Controller, Get, Param } from '@nestjs/common';
+
 import { CloudConfigService } from './cloud-config.service';
 
 @Controller(':configGroup')
