+++ src/modules/cloud-config/constants.ts
@@ -1,1 +1,1 @@
-export const CONFIGS_MAP_KEY = Symbol('CONFIGS_MAP_KEY');
+export const CONFIG_MAP_KEY = Symbol('CONFIGS_MAP_KEY');
